pyproject.toml
[tool.poetry]
name = "spotify-visualization-project"
version = "0.1.0"
description = ""
authors = ["necabotheking <aichacamara@uchicago.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
black = "^23.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
